--- a/Cargo.toml
+++ b/Cargo.toml
@@ -52,9 +52,8 @@ version = "0.5.0"
 features = ["serde_support"]
 git = "https://github.com/KZen-networks/rust-gmp"
 
-[dependencies.secp256k1]
-version = "0.12.0"
-features = ["rand", "serde"]
+[dependencies.libsecp256k1-rs]
+version = "0.2.4"
 
 [dependencies.curve25519-dalek]
 version = "1.0.1"