--- a/src/elliptic/curves/mod.rs
+++ b/src/elliptic/curves/mod.rs
@@ -11,7 +11,7 @@ extern crate cryptoxide;
 extern crate curve25519_dalek;
 extern crate pairing;
 extern crate sapling_crypto;
-extern crate secp256k1;
+extern crate libsecp256k1_rs;
 
 pub mod curve_jubjub;
 pub mod curve_ristretto;