--- a/src/elliptic/curves/mod.rs
+++ b/src/elliptic/curves/mod.rs
@@ -9,9 +9,9 @@ extern crate rand;
 
 extern crate cryptoxide;
 extern crate curve25519_dalek;
+extern crate libsecp256k1_rs;
 extern crate pairing;
 extern crate sapling_crypto;
-extern crate libsecp256k1_rs;
 
 pub mod curve_jubjub;
 pub mod curve_ristretto;