#43
test passed by `cargo test --features ec_secp256k1`

type                    | before    | after   | test code link |
------------------|-----------|--------|-------|
1 of 1 aggregate | 112ms   |   3ms  |  [aggregate_1_of_1.py](https://github.com/namuyan/multi-party-schnorr/blob/master/aggregate_1_of_1.py) |
n of n aggregate | 458ms   |   7ms  | [aggregate_n_of_n.py](https://github.com/namuyan/multi-party-schnorr/blob/master/aggregate_n_of_n.py)  |
t of n threshold | 2475ms   |  68ms |  [thresholdbig_t_of_n.py](https://github.com/namuyan/multi-party-schnorr/blob/master/thresholdbig_t_of_n.py) |

performance improved!
please check my code, thank you.