@namuyan 2019-07-03
Now I find two library [libsecp256k1-rs](https://source.that.world/source/libsecp256k1-rs/) and [libsecp256k1](https://github.com/tari-labs/libsecp256k1).
**libsecp256k1-rs** is fork of **libsecp256k1**, and have Copy trait and overwrite operators but is not main stream and no production code. 
**libsecp256k1** require wide modification and has influence to others, how do you think?