@omershlo 2019-07-03
Hi @namuyan - looks like great work. 
We started experimenting with  this idea in another project: https://github.com/KZen-networks/emerald-city 
there we took only the secp256k1 relevant parts of curv and replaced them with  libsecp256k1 (the second lib, the native one).