@namuyan 2019-07-03
Thank you for reply.
I now use **curv** for my project, but **emerald-city** looks more functional and pure-Rust implement (and compatibility with curv?).
I may use the latter, thank you.