@namuyan 2019-07-04
I just change README.md? 
UNLUCKY DICE, It looks a low repeatability bug.
https://github.com/KZen-networks/curv/blob/master/src/cryptographic_primitives/commitments/hash_commitment.rs#L78