@omershlo 2019-07-04
you are right, we just need to run it again. The problem is that it is unknown for me how to do unit test for probabilistic code. We got stuck on this a while ago (https://github.com/KZen-networks/curv/issues/10). 

Any suggestions will be welcome.