@vhnatyk 2019-07-07
> Thank you for reply.
> I now use **curv** for my project, but **emerald-city** looks more functional and pure-Rust implement (and compatibility with curv?).
> I may use the latter, thank you.

https://github.com/rust-bitcoin/rust-secp256k1/pull/125#issuecomment-508598506

According to latest merged PR in original [rust-secp256k1](https://github.com/rust-bitcoin/rust-secp256k1) - it now supports WASM that was only benefit of [libsecp256k1](https://github.com/tari-labs/libsecp256k1). Performance benefit over **rust-secp256k1** is supposedly due to later being constant time. Additionally **rust-secp256k1** is a wrapper over original Bitcoin's **secp256k1** that is most tested and commonly used library out there.