@omershlo 2019-07-07
@vhnatyk I agree. Let's switch back.
 @elichai kudos