@elichai 2019-07-07
Glad to hear :)
Using good vetted crypto libraries is important :)