@elichai 2019-07-23
FYI After this: https://github.com/rust-bitcoin/rust-secp256k1/pull/132
There should be a ~15% performance improvement in verification time.