@elichai 2019-07-23
Oh.
I just realized why your performance is so bad.
you launch a new context for every operation: https://github.com/KZen-networks/emerald-city/pull/6/files#diff-cd37acb447aaca91b67e2e383e1b43afL391

Context creation in Secp256k1 is the most intensive part of the library.

You should use `lazy_static!` or `std::syn::Once()`(example: https://github.com/elichai/ecc-secp256k1/blob/master/src/secp256k1.rs#L480)
to create the context *once* and then use it everywhere.
(the context is thread safe as it doesn't mutate after creation)