@elichai 2019-07-23
Ok. let's put the benchmarking stats to bed once and for all:
I used this repo: https://github.com/namuyan/multi-party-schnorr on commit: https://github.com/namuyan/multi-party-schnorr/commit/6c11c408a064225b990ec55bbb291994791031b8

Before I changed anything I got consistently these stats:
1-of-1 = 142 mSec
n-of-n = 447 mSec
t of n threshold = 3048 mSec

After I changed the context to be created *before* the tests and then used it globally I got these stats:
1-of-1 = 1mSec
n-of-n = 1.5 mSec
t of n threshold = 42mSec

And this is with *a lot* of over head.
in secp256k1 sign/verify time are measured in uSec(micro). not mSec(mili).

@omershlo @namuyan.
If you want I can open a PR for lazy_static/sync::Once