@elichai 2019-07-23
Will do.
Do you prefer unsafe code or a new dependency? (in reality their doing almost the same thing)