@namuyan 2019-07-24
Thank you for quality up!

> After I changed the context to be created before the tests and then used it globally I got these stats:
1-of-1 = 1mSec
n-of-n = 1.5 mSec
t of n threshold = 42mSec

Is this performance with emerald-city? or curv?