@elichai 2019-07-25
curv. and create a key before you measure performance so that the context will be available already (you shouldn't count the context creation as part of the performance test)